hotel_booking_service/pyproject.toml
[project]
name = "hotel-booking-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Vlad Zavorotnyy", email = "vlad.zavorotnyy@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "pre-commit (>=4.1.0,<5.0.0)",
    "ruff (>=0.9.10,<0.10.0)",
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]

[tool.poetry]
packages = [{include = "hotel_booking_service", from = "src"}]

[tool.poetry.scripts]
manage = "src.manage:main"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.1.0"
pytest-django = "^4.10.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
